requirements.txt
Adafruit-Blinka==5.3.4
adafruit-circuitpython-dht==3.5.1
Adafruit-PlatformDetect==2.16
Adafruit-PureIO==1.1.5
astroid==2.4.2
autopep8==1.5.4
AWSIoTPythonSDK==1.4.9
colorzero==1.1
gpiozero==1.5.1
isort==5.5.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pyftdi==0.51.2
pylint==2.6.0
pyserial==3.4
python-dotenv==0.14.0
pyusb==1.0.2
rpi-ws281x==4.2.4
RPi.GPIO==0.7.0
six==1.15.0
sysv-ipc==1.0.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
